+++ NEWS.md
@@ -16,6 +16,25 @@ And constructed with the following guidelines:
   (and resets the patch)
 * Bug fixes and misc changes bumps the patch
 
+
+ <a href="https://github.com/trinker/qdapRegex" target="_blank">qdapRegex</a> 0.3.3
+----------------------------------------------------------------
+
+**BUG FIXES**
+
+**NEW FEATURES**
+
+* `as_numeric` added for use with `rm_number`.  This is a wrapper for 
+  `as.numeric(gsub(",", "", x))`, which removes commas and converts a vector of 
+  strings to numeric.
+
+**MINOR FEATURES**
+
+**IMPROVEMENTS**
+
+**CHANGES**
+
+
  <a href="https://github.com/trinker/qdapRegex" target="_blank">qdapRegex</a> 0.2.1 - 0.3.2
 ----------------------------------------------------------------
 
@@ -75,7 +94,6 @@ And constructed with the following guidelines:
 * `rm_caps_phrase` has a new regular expression that is more accurate and does 
   not pull trailing white space.
 
-**CHANGES**
 
  <a href="https://github.com/trinker/qdapRegex" target="_blank">qdapRegex</a> 0.1.3 - 0.2.0
 ----------------------------------------------------------------